scripts/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Script to upload HN comments to HuggingFace dataset"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.2.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
]
